add test data and create separate file, init.py, for initialising the database
